refactor: relocate component to containers
